Issues with implementation based on rclpy, temporal fix
This commit proposes a quick fix to cope with the limitations
identified in both OS X and Linux. Essentially, calls to fetch nodes
topics and services deliver incomplete information. This commit makes
several calls to such services and keeps the one that delivers the most
complete information.

It should be noted that this piece of code has been developed assuming that
network status (Nodes/topics/services) won't change during the time of
sampling